Move Message components into their own modules
