+++ src/views/Messages/Bubble.js
@@ -1,0 +1,22 @@
+import React, { PropTypes } from 'react';
+import { View } from 'react-native';
+import styles from './styles';
+
+
+const Bubble = ({
+  children,
+  style,
+}) => (
+  <View style={[styles.bubble, style]}>
+    {children}
+  </View>
+);
+
+
+Bubble.propTypes = {
+  children: PropTypes.any,
+  style: PropTypes.any,
+};
+
+
+export default Bubble;
