+++ src/views/Messages/index.js
@@ -1,18 +1,10 @@
-import { fromPairs } from 'lodash';
-import React, { Component, PropTypes } from 'react';
+import React, { PropTypes } from 'react';
+import { View, ScrollView } from 'react-native';
 
-import {
-  View,
-  ScrollView,
-  Image,
-  TouchableWithoutFeedback,
-  TextInput,
-} from 'react-native';
-
-import { BaseView, PatternBackground, Text } from 'src/components';
-import images from 'src/constants/images';
-import * as constants from 'src/constants/messages';
+import { BaseView, PatternBackground } from 'src/components';
 import styles from './styles';
+import MessageGroup from './MessageGroup';
+import Send from './Send';
 
 
 const Messages = ({
@@ -33,129 +25,8 @@ const Messages = ({
 );
 
 
-const MessageGroup = ({
-  messages,
-}) => (
-  <View>
-    {messages.map((message, i) => <Message key={i} {...message} />)}
-  </View>
-);
-
-
-const messageTypeStyles = fromPairs([
-  [constants.MESSAGE_DIRECTION_OUTBOUND, {
-    bubble: styles.bubbleOutbound,
-    messageContent: styles.messageContentOutbound,
-  }],
-  [constants.MESSAGE_DIRECTION_INBOUND, {
-    bubble: styles.bubbleInbound,
-    messageContent: styles.messageContentInbound,
-  }],
-]);
-
-
-const Message = ({
-  details: { direction },
-  content,
-}) => {
-  const directionStyles = messageTypeStyles[direction];
-
-  return (
-    <View style={styles.message}>
-      <Bubble style={directionStyles.bubble}>
-        <Text style={[styles.messageContent, directionStyles.messageContent]}>
-          {content}
-        </Text>
-      </Bubble>
-    </View>
-  );
-};
-
-
-class Send extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { content: '' };
-    this.onSendPress = this.onSendPress.bind(this);
-    this.onTextChange = this.onTextChange.bind(this);
-  }
-
-  onTextChange(content) {
-    this.setState({ content });
-  }
-
-  onSendPress() {
-    this.props.onSendPress({ content: this.state.content });
-  }
-
-  render() {
-    return (
-      <View style={styles.send}>
-        <Bubble style={styles.bubbleSend}>
-          <TextInput
-            multiline
-            uid="sendInput"
-            placeholder="Write something"
-            placeholderTextColor={images.SEND_INPUT_PLACEHOLDER_TEXT}
-            underlineColorAndroid="transparent"
-            style={styles.sendInputText}
-            onTextChange={this.onTextChange}
-          />
-        </Bubble>
-
-        <TouchableWithoutFeedback
-          uid="sendButton"
-          onPress={this.onSendPress}
-        >
-          <View style={styles.sendButton}>
-            <Image
-              style={styles.sendButtonImage}
-              source={images.SEND_MESSAGE_ICON}
-            />
-          </View>
-        </TouchableWithoutFeedback>
-      </View>
-    );
-  }
-}
-
-
-const Bubble = ({
-  children,
-  style,
-}) => (
-  <View style={[styles.bubble, style]}>
-    {children}
-  </View>
-);
-
-
 Messages.propTypes = {
   groups: PropTypes.arrayOf(PropTypes.object),
-  onSendPress: PropTypes.func.isRequired,
-};
-
-
-MessageGroup.propTypes = {
-  messages: PropTypes.arrayOf(PropTypes.object),
-};
-
-
-Bubble.propTypes = {
-  children: PropTypes.any,
-  style: PropTypes.any,
-};
-
-
-Message.propTypes = {
-  details: PropTypes.shape({
-    direction: PropTypes.string.isRequired,
-  }).isRequired,
-  content: PropTypes.string,
-};
-
-
-Send.propTypes = {
   onSendPress: PropTypes.func.isRequired,
 };
